vinyl: fix 0 division in case of canceled dump
It can happen, that vinyl index dump is scheduled, but the index
is dropped before the dump starts. In this case the dump task is
finished with 0 dump time. The time then is used in a log message
to obtain dump rate as byte count / dump time. Since dump time is
0, this is undefined behaviour. Moreover, it was not even a dump,
so nothing to log here.

Also dump time is used to update throttling parameters in
vy_regulator. Here if the dump took 0 time, there was no a dump,
so it is just ignored.

Part of #4609

(cherry picked from commit cc7da0e57d947eac17eb6779a30deddc1918e460)