test: remove skip condition of app-tap/pwd test
During creating patch:

  e3d9d8c97edd39438cbc98fccac996e61ce4984c ("build: add CentOS 8 into CI / CD")

was found the issue:

  [001] app-tap/pwd.test.lua                                            [ fail ]
  [001] Test failed! Output from reject file app-tap/pwd.reject:
  [001] TAP version 13
  [001] 1..6
  [001] ok - checking user by id
  [001] ok - checking user by name
  [001] ok - checking group by id
  [001] ok - checking group by name
  [001]
  [001] Last 15 lines of Tarantool Log file [Instance "app_server"][/mnt/test/var/001_app-tap/pwd.test.lua.tarantool.log]:
  [001] tarantool: /lib64/libcurl.so.4: no version information available (required by tarantool)
  [001] builtin/pwd.lua:169: getpwall failed [errno 2]: No such file or directory

Currently it was tried to reproduce the issue, but all the attempts
didn't got the needed fail. After discussion of the issue decided to
close the current one as not the issue for the test and continue the
fixes for getpwall/getgrall functions within issue #5034.

Closes #4592

(cherry picked from commit b92549ad24f356a5a9a5db82964cdb148d17455d)