test: update test-run
Introduce the new test-run.py --update-result option, which allows to
write a new result file (a test reference output) or re-write an
existing one.

Before this patch, test-run did not require the option to write a new
result file (when there is no exising one). Now 'fail' status will be
reported for such tests when the option is not passed.

Closes https://github.com/tarantool/test-run/issues/194