vinyl: fix squashing set and arithmetic operations
Before this patch in upsert squashing procedure (which is implemented as
xrow_upsert_squash() function) was missed the case when first (old one)
operation is set and second one is addition or subtraction. This led to
the wrong result of squashing these operations. Let's fix it and handle
such case: {'+', x} applied on the top of {'=', y} is {'=', x+y}.

Closes #5106
Part of #5107

(cherry picked from commit a3de4f742314bb3ff2bb0aa881f9ae4a62467019)