build: update curl submodule to 7.71.1 version
The submodule was on 7.66.0 version. This update closes security
problem for us:

May leak a part of a password in a DNS query when the password
contains '@' symbol. It occurs on an HTTP redirect, so one of
possible workarounds is to disable follow_location option
(available since 2.1.2-42-gdf87154ff within 2.1 series and since
2.2.0-202-gdaf1ced88 in 2.2+ series). Another way: verify that a
password does not contain the '@' symbol.

https://curl.haxx.se/docs/CVE-2020-8169.html

See also release notes:

https://curl.haxx.se/changes.html#7_71_1
https://curl.haxx.se/changes.html#7_71_0
https://curl.haxx.se/changes.html#7_70_0
https://curl.haxx.se/changes.html#7_69_1
https://curl.haxx.se/changes.html#7_69_0
https://curl.haxx.se/changes.html#7_68_0
https://curl.haxx.se/changes.html#7_67_0

Closes #4698

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>