small: sanitized rlist, bug in lsregion, and new region API
Backport of commit ffec9e40d8695bffe76109b721b1050db84b384e,
I've to update all region_alloc_array calls to use fuber()->gc
as there was no txn->region yet.

Rlist used a hack to implement offsetof() leading to crash under
undefined behaviour clang sanitizer. It was fixed in this update.

Additionally, region_alloc_object() is changed to return the used
size and a new macro region_alloc_array() is added. This small
API change is supposed to simplify switching lots of region
allocations to aligned versions in scope of #4609.

And finally - there was a bug in lsregion, when allocation was
exactly of slab size. It is fixed in small's master.

Part-of #4609
Part-of #5060

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>