xlog: introduce an error code for XlogGapError
XlogGapError object didn't have a code in ClientError code space.
Because of that it was not possible to handle the gap error
together with client errors in some switch-case statement.

Now the gap error has a code.

This is going to be used in applier code to handle XlogGapError
among other errors using its code instead of RTTI.

Needed for #5287

(cherry picked from commit fc8e22972275272cf828dc783d148c16152af6b5)